Update Encodable/Decodable APIs
Now there's an explicit contract between Encodable/Decodable. If a type
implement Encodable then it must be possible to decode back to the
original format.

Using type parameters instead of associated types also has the advantage
of being able to define multiple implementations related to a type. For
example `u64` and `[f32; 2]` will probably encode into the same format,
for example `[u8; 9]`. This was not possible with the previous
implementation

The `empty` methods were removed from the Encodable/Decodable traits and
moved to a shared subtrait `Init`. Init serves the same purpose as
`Default` from the standard library but can be implemented without
restrictions.