refactor(tensor_ops): Simplify fault injection glue code.
It no longer keeps track of devices because as far as I can tell using
copy_ doesn't change the device of the original, just the data.

`.numpy()` calls were changed to force. I'm not sure why I thought it
necessary to not use it before. The original remains unchanged and the
numpy one will result in a copy to the cpu memory.