fix(EncodedSystem): Fix caching for decoded tensors
This cannot be done inside EncodedSystem because there is no guarantee
about which instance of `data` is given as an input to any of the
functions. This means that there is no reliable way to reset the cache.
If we move the cache inside the specific implementation however then
it's always sound because that's the actual thing that gets copied.